refactor redis settings
