+++ core/api_clients/magic_rust/models.py
