--- core/clients/server_data_api/__init__.py
@@ -1,2 +1,0 @@
-from .api import MagicRustServerDataAPI
-from .models import FullServerData, MonitoringServerData, ServerTypes
